v4l: xilinx: hdmirx: Add frl and video clock properties
commit  3446853d571dcde6d66152f97b946bc59a3b3242 from
https://github.com/Xilinx/linux-xlnx.git xlnx_rebase_v5.10

This patch adds the FRL and Video clock frequency properties to the
driver. These are used to calculate the frame rate and program the vcke
ratio in FRL mode. The clock frequencies are to be mentioned in KHz.

Signed-off-by: Vishal Sagar <vishal.sagar@xilinx.com>
State: pending
Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>