s32: adc: Correctly check bus clock frequency based on S32 platform flavours
commit 0b953d84bda5e05bc061c978a819a55d15b6d9ce from
https://source.codeaurora.org/external/autobsps32/linux

On S32V234 platforms, using the 'ADCLKDIV' field from the MCR register,
there is the posibility of dividing the bus clock frequency by four.
On the other hand, on S32-GEN1 platform, this field is reserved,
and therefore the posibilities are to directly use the bus clock
frequency or divide it by two, by using the 'ADCLKSEL' field from
the same MCR register.

Issue: ALB-7726
Signed-off-by: Ciprian Marian Costea <ciprianmarian.costea@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>