regulator: rpi-panel: Remove get_brightness hook
commit  ce646f7446fba781ee55ea64997b9ad3f6bd8564 from
https://github.com/raspberrypi/linux.git rpi-5.10.y

The driver was implementing a get_brightness function that
tried to read back the PWM setting of the display to report
as the current brightness.
The controller on the display does not support that, therefore
we end up reporting a brightness of 0, and that confuses
systemd's backlight service.

Remove the hook so that the framework returns the current
brightness automatically.

Signed-off-by: Dave Stevenson <dave.stevenson@raspberrypi.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>