octeontx2-af: Do not fixup all VF action entries
commit 15d79e8350e35548a27244dde8579321cce51127 from
git@git.assembla.com:cavium/WindRiver.linux.git

AF modifies all the rules destined for VF to use
the action same as default RSS action. This fixup
was needed because AF only installs default rules with
RSS action. But the action in rules installed by a PF
for its VFs should not be changed by this fixup.
This is because action can be drop or direct to
queue as specified by user(ntuple filters).
This patch fixes that problem.

Fixes: d1f32fa5c67b ("octeontx2-af: Update forwarding rule action targeting VF")
Change-Id: Id286d362df2979124df8bc8cbffca11fc33156a8
Signed-off-by: Subbaraya Sundeep <sbhatta@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/62119
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>