net: axienet: Properly release all tx resources when done
When adding support for mcdma the free_tx_chain function was
accidentally broken. The loop that scans the tx descriptors
for completed buffers does not increment the consumer index
tx_bd_ci, that is done by the caller in bulk, so the check
for completed buffers always failed on the second iteration
since it was looking at the previously cleared status bit.

This manifests itself in a very strange way, programs that
send many large frames in row can completely eat up the
tx descriptors and they will never be returned and the skb
buffers never reclaimed. The process was always hanging waiting
for "wmem" in sock_alloc_send_pskb()

root@machine:/proc/3880# cat stack
[<0>] __switch_to+0x10c/0x170
[<0>] sock_alloc_send_pskb+0x160/0x270
[<0>] packet_sendmsg+0x810/0x132c
[<0>] __sys_sendto+0xdc/0x150
[<0>] __arm64_sys_sendto+0x2c/0x3c
[<0>] el0_svc_common.constprop.0+0x78/0x1c4
[<0>] do_el0_svc+0x24/0x8c
[<0>] el0_svc+0x14/0x20
[<0>] el0_sync_handler+0x1a4/0x1b0
[<0>] el0_sync+0x178/0x180

Fixes: c606f22e43918 (net: axienet: added multichannel DMA support)
Signed-off-by: Ivar Holmqvist <ivar.holmqvist@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>