media: ti: j721e-csi2rx: Drain DMA when dropping frames
commit 01499f804027ecfdf8e6ddc081a6c0b52c91518b from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

Once frames start dropping due to lack of buffers, some data gets stuck
in the DMA pipeline somewhere. So the first DMA transfer after frame
drops gives a partial frame. This is obviously not useful to the
application and will only serve to confuse it. Issue a DMA transaction
to drain that up so that the next frame comes out clean.

Signed-off-by: Pratyush Yadav <p.yadav@ti.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>