dt-bindings: memory-controllers: Introduce ti,gpmc-child
commit ee11f00660eda4f1f8392a5ee17ee20b30211c83 from
git://git.ti.com/ti-linux-kernel/ti-linux-kernel.git

commit 04f461f35e631da4a4c4b8d1a02cc35aa80dc161 upstream.

This binding is meant for the child nodes of the TI GPMC node. The node
represents any device connected to the GPMC bus. It may be a Flash chip,
RAM chip or Ethernet controller, etc. These properties are meant for
configuring the GPMC settings/timings and will accompany the bindings
supported by the respective device.

Signed-off-by: Roger Quadros <rogerq@kernel.org>
Signed-off-by: Tony Lindgren <tony@atomide.com>
Signed-off-by: Vignesh Raghavendra <vigneshr@ti.com>
Reviewed-by: Rob Herring <robh@kernel.org>
Signed-off-by: Xulin Sun <xulin.sun@windriver.com>