arm64: zynqmp: adjust flash size for zcu102 rev1.0 board
There are mt25qu512a(64MB)x2 flashes at zcu102 rev1.0 board. So we add
an overlay partition node to enlarge the total flash size to 128MB.

And this is also for flash_lock/unlock since the top sectors in flash
must be visible in order to be locked/unlocked.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>