arm64: zynqmp: adjust qspi flash partition
There are 16MBx2 flashes at zcu102 revA board, so adjust the total size
of partitions to 32MB. This is also for flash_lock/unlock since the top
sectors in flash must be visible in order to be locked/unlocked.

And adjust the order and size of partitions to make sure that the start
address of partitions is Megabyte in power of 2. This adjustment is for
convenience to test flash_lock/unlock since the sectors to be protected
are always as a power of 2.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>