octeontx2-af: fix detection of IP layer
commit 7051287fbd2d7ffe57b28dfa1bea375f97daed0d from
git@git.assembla.com:cavium/WindRiver.linux.git

Checksum and length checks are not enabled for IPv4 header with
options and IPv6 with extension headers.
To fix this a change in enum npc_kpu_lc_ltype is required which will
allow adjustment of LTYPE_MASK to detect all types of IP headers.

Change-Id: Ic89492528a821e3c05f5524fcb5d00d4708da034
Signed-off-by: Michal Mazur <mmazur2@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/22262
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <Sunil.Goutham@cavium.com>
[RQ: Drop the parts which have been already applied.]
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>