Revert "mtd: spi-nor: Simplify odd address handling in read for dual parallel mode"
This reverts commit 634a317469fe7d6b59d41d5cf5b59fc827864741.

The commit ("mtd: spi-nor: Simplify odd address handling in read for
dual parallel mode") introduces an issue that passes an odd address to
spi driver. When "nor->isparallel && (from & 1)" is true, running "++buf"
will force the buffer to be an odd address and then pass it to the
function zynqmp_qspi_exec_op as op->data.buf.in. The spi driver using this
odd address to do DMA operations will receive the corrupted data.

Signed-off-by: Quanyang Wang <quanyang.wang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>