octeontx2-af: fix issue with returning vf base steering rule
commit 41f63d85129a58972824325023a3b4ece568c478 from
git@git.assembla.com:cavium/WindRiver.linux.git

When pf netdev installs traffic steering rule to vf, the vf base
steering rule is not saved. As a consequence, when a requester reads
the base steering rule, the default dmac rule is returned, and all
further traffic steering rules are based on dmac, which is not the
expected functionality. This patch fixes the issue by saving the vf
base steering rule for pf netdev installed rules. When vf requester
reads base steering rule, the high priority rule is returned to vf.

Fixes: 15d79e8350e3 ("octeontx2-af: Do not fixup all VF action entries")
Change-Id: I83abdd86cec1564e6e40c73e322198f9bce0119f
Signed-off-by: Naveen Mamindlapalli <naveenm@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/64361
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>