HSD #1308655004: don't clear the ue/ce status bits
commit  d60212891e37be9f599cc3655393e1d1095221fc from
https://github.com/altera-opensource/linux-socfpga.git

When there is a uncorrected or double-bit error, the kernel will panic
and will result in a watchdog reset. When the SDM gets the reset, it
needs to know that a DBE had occurred so it can scrub the DDR on a
reboot. Thus we should not clear the CE/UE status bits in the EDAC.

Signed-off-by: Dinh Nguyen <dinh.nguyen@intel.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>