soc: marvell: cn10k SerDes: fix stopTraining() call
commit 014c4e94241b6f1f4e76d4992f15ce4ac652ded0 from
git@git.assembla.com:cavium/WindRiver.linux.git

This fixes the case, when portm is in multilane mode
and the rx training was started only for the specific
lane (but other than zero) and the training happens to
fail due to timeout (does not complete before the
timeout). In that case the stopTraining() would be
requested for the wrong lane number.

Signed-off-by: Damian Eppel <deppel@marvell.com>
Change-Id: Iec4fdf86ff02008fa746d80d0dbfb18b6e007658
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/64911
Reviewed-by: Chandrakala Chavva <cchavva@marvell.com>
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>