HSD #1508914759: firmware: stratix10-svc: fix FPGA reconfiguration SVC timeout
commit  d9582455fa7e881859850b2bbd0bd1896e016b4a from
https://github.com/altera-opensource/linux-socfpga.git

Modifying the driver to return a single callback for every SMC call.
As the SVC driver sends a callback for every block of memory reclaimed at the end of
every FPGA data transfer, it triggers multiple callbacks.
The additional callbacks triggers the FPGA driver to start a new block of 512kBytes
transfer without waiting for the previous 512kBytes of data to be done.
As a result, it starts waiting earlier and eventually time-outs.
The change also  increases the time-out of the FPGA reconfiguration
as it is too close to the threshold.

Reported-by: Russell Zhang <russel.zhang@nokia-sbell.com>
Signed-off-by: Ang Tien Sung <tien.sung.ang@intel.com>
Signed-off-by: Meng Li <Meng.Li@windriver.com>