arm64: Restore the init of srcin for __arch_copy_{from,to}_user()
When we were trying to merge the v5.10.77 stable kernel, the commit
d4d9c065988c ("arm64: Avoid premature usercopy failure") conflicts with
the changes in the Marvell specific workaround commit 8b8b3bc1497a
("arm64: Use nops between LDs and STs"). In the fixup of these
conflicts, the init of srcin was missed in the function
__arch_copy_{from,to}_user(). Restore these codes back to fix the kernel
hang when running the fio testcase.

Signed-off-by: Kevin Hao <kexin.hao@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>