s32gen1: qspi: suppress warning using __force
commit 6f4d8d00eff5f16843a40d456c1924d8c99e2483 from
https://source.codeaurora.org/external/autobsps32/linux

The __invalid_dcache_area() required as the first argument a 'void *'
but ahb_addr is of type 'void __iomem *'. This generates a warning
in sparse. The possible options were to either cast it to 'void *'
that will generate another warning from sparse that __iomem was removed
during casting or to use __force to suppress the sparse warning

Issue: ALB-8246
Signed-off-by: Andrei Botila <andrei.botila@nxp.com>
Signed-off-by: Zhantao Tang <zhantao.tang@windriver.com>