arm64: psci: don't create online interface file for cpu0 on intel socfpga platform
After upgrading u-boot to v2021.07, it creates the fit bootloader image
that includes atf and u-boot by default. As known, the cpu hardware
off/on action is processed in atf code, but for intel socfpga platform,
system resets when do cpu0 up. So, create this workaround in kernel
code to avoid operating cpu0 off/on interface online file firstly, and
then try to fix the root cause in atf code in later.

Signed-off-by: Meng Li <Meng.Li@windriver.com>