pstore/ftrace: Add and use ftrace_test_recursion_trylock_safe
We experienced system hang when using ramoops/ftrace.

To avoid the deadlock causing the hang, this patch adds a safe version of
ftrace_test_recursion_trylock. This version does not allow any recursion
at the same context level (normal, softirq, irq and NMI) to avoid potential
deadlock. This is based on the Steven Rostedt's patch(not yet in mainline)
in following link.

Link: https://lkml.org/lkml/2021/6/10/868

Signed-off-by: He Zhe <zhe.he@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>