Revert "scripts: Makefile: Enable creation of _symbols_ DT node for overlays"
commit 3c02b480dc20f5f65afc7d51625b5fda40525672 from
https://source.codeaurora.org/external/imx/linux-imx

This reverts commit 11db97cab885454945c4e36d86a27864e994bad3.

The blamed commit changes the common Makefile and force creation of
__symbols__ node on the generated dtb files for all platforms.

This is not good and not acceptible under some cases due to this change
will increase the final DTB size a lot and bring big impact for others
who does not need creation of _symbols_ DT node for overlays.For
example, on i.MX OP-TEE, the maxsize of DT is 1MB, this patch will break
some of the i.MX6/i.MX7 boards to boot with OP-TEE enabled.

A similar patch was submitted and rejected by the community. The
following post discusses this:

https://lore.kernel.org/patchwork/patch/821645/

LS1028A-QDS, which needs the "-@" build flag for device tree overlays,
now sets the "DTC_FLAGS_fsl-ls1028a-qds" GNU Make variable in
arch/arm64/boot/dts/freescale/Makefile, and the support for overlays is
already built for it, so the reverted patch was not even necessary.

Signed-off-by: Jason Liu <jason.hui.liu@nxp.com>
Signed-off-by: Vladimir Oltean <vladimir.oltean@nxp.com>
Signed-off-by: Xiaolei Wang <xiaolei.wang@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>