PCI: octeontx2-pem-ep: Add PERST interrupt handling
commit c2ff3d8056fb199fe98b6af2f1fa101a1edc812f from
git@git.assembla.com:cavium/WindRiver.linux.git

In EP mode the PERST interrupt is generated when host resets. This patch
adds support to handle the interrupt. UIO support is added for it to
help user space application to receive the PERST interrupt event in case
we need to do anything from userspace.

Signed-off-by: Radha Mohan Chintakuntla <radhac@marvell.com>
Change-Id: I62106a47ae682169b99991762ede9dd042146e9c
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/73209
Tested-by: sa_ip-sw-jenkins <sa_ip-sw-jenkins@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
[RQ: Add select UIO for dependence.]
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>