octeontx2-af: cn10k: fix adding large offset to ptp clock
commit 7a7845b9dce68471e25279cda1008e32927a69ec from
git@git.assembla.com:cavium/WindRiver.linux.git

The negative delta adjustment in ptp_adj_clock() is causing
ptp clock to jump to a large value because the result of adding
delta to nsecs is negative, and it is assigned to an unsigned
variable. This patch fixes the issue by adding the delta to
entire timestamp (secs+nsecs) instead of just nsec counter value.

Fixes: 417dc51aa75f ("octeontx2-pf: Use HW registers in PTP instead
                      of timercounter abstraction")
Change-Id: Ia2daff72817da890a0ad563068aeb12cde0ce463
Signed-off-by: Naveen Mamindlapalli <naveenm@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/74840
Reviewed-by: Hariprasad Kelam <hkelam@marvell.com>
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>