octeontx2-af: Fix key checking for source mac
commit 3871bb6fde0d7f3cd6da4893685c526e73d3bec6 from
git@git.assembla.com:cavium/WindRiver.linux.git

Given a field with its location/offset in input packet,
the key checking logic verifies whether extracting the
field can be supported or not based on the mkex profile
loaded in hardware. This logic is wrong wrt source mac
and this patch fixes that.

Fixes: 614f9226127e ("octeontx2-af: Support installing TX MCAM rules")
Change-Id: I43cdb789f564d101ef374186a427c4d5f9e192a1
Signed-off-by: Subbaraya Sundeep <sbhatta@marvell.com>
Reviewed-on: https://sj1git1.cavium.com/c/IP/SW/kernel/linux/+/74554
Reviewed-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Tested-by: Sunil Kovvuri Goutham <sgoutham@marvell.com>
Signed-off-by: Ruiqiang Hao <Ruiqiang.Hao@windriver.com>