trace_events_trigger: fixup for v5.10.113 backport on -rt
The v5.10.113 brought a backport of commit ce33c845b030 "tracing:
Dump stacktrace trigger to the corresponding instance" and in that
backport it was switched from the three arg instance of __trace_stack()
to the 4 arg instance which was necessary for the older v5.10 baseline.

However the v5.10-rt has a pre-mainline snapshot of the trace reduction
args;  commit 36590c50b2d0729952511129916beeea30d31d81 "tracing: Merge
irqflags + preempt counter." -- so it in fact needs the non-converted
mainline backport.  So switch back to the 3 arg instance to solve:

  kernel/trace/trace_events_trigger.c:1227:17: error: too many arguments
  to function '__trace_stack'

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>