regulator: consumer: Add missing stubs to regulator/consumer.h
commit 51dfb6ca3728bd0a0a3c23776a12d2a15a1d2457 upstream

Add missing stubs to regulator/consumer.h in order to fix COMPILE_TEST
of the kernel. In particular this should fix compile-testing of OPP core
because of a missing stub for regulator_sync_voltage().

Reported-by: kernel test robot <lkp@intel.com>
Signed-off-by: Dmitry Osipenko <digetx@gmail.com>
Link: https://lore.kernel.org/r/20210120205844.12658-1-digetx@gmail.com
Signed-off-by: Mark Brown <broonie@kernel.org>

Backport from mainline to fix build failure when CONFIG_REGULATOR=n which
is default in linux-yocto.
5.10.114 in mainline stable tree does not handle this case.
Link: https://lore.kernel.org/all/62796b82.1c69fb81.1bf47.0bcf@mx.google.com/

Signed-off-by: He Zhe <zhe.he@windriver.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@gmail.com>